feat: 메인 페이지 정비
